Feat: 채팅 삭제 기능 수정 및 콘텐츠 추천 데이터 정제 + favicon 추가

+++ web/chat/views.py
@@ -1,11 +1,11 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.utils import timezone
+from chat.utils import get_image_response
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST, require_http_methods
 from django.http import JsonResponse, FileResponse, HttpResponseNotAllowed, Http404, HttpResponseNotFound, HttpResponseServerError, HttpResponse
 from user.models import User
 from .models import Chat, Message, Content, MessageImage, UserReview
-import pandas as pd
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.metrics.pairwise import cosine_similarity
 from dogs.models import DogProfile, DogBreed
@@ -17,6 +17,7 @@ from datetime import date, timedelta
 import uuid
 import requests
 import json
+import pandas as pd
 import base64
 from django.template.loader import render_to_string, get_template
 import os
@@ -468,18 +469,19 @@ def chat_send(request):
 
 @require_POST
 @csrf_exempt
-def chat_member_delete(request, chat_id):
+def chat_member_delete(request, dog_id, chat_id):
     try:
-        chat = get_object_or_404(Chat, id=chat_id)
+        if request.method == "POST":
+            chat = get_object_or_404(Chat, id=chat_id, dog_id=dog_id)
 
-        if not is_chat_owner(request, chat):
-            return JsonResponse({'status': 'unauthorized'}, status=403)
+            if not is_chat_owner(request, chat):
+                return JsonResponse({'status': 'unauthorized'}, status=403)
 
-        chat.delete()
-        return JsonResponse({'status': 'ok'})
+            chat.delete()
+            return JsonResponse({'status': 'ok'})
 
     except Chat.DoesNotExist:
-        return JsonResponse({'status': 'not_found'}, status=404)
+        return JsonResponse({'error': 'Invalid method'}, status=405)
     
 
 @require_POST
@@ -673,7 +675,6 @@ def recommend_content(request, chat_id):
         "cards_html": html,
         "has_recommendation": True
     })
-
 
 @csrf_exempt
 @require_POST
